Moved common functions from insert.ts to tree.ts - `recomputeTreeMetadata` - `calculateCharCount` - `calculateLineFeedCount`
